1.06 [fast]
Changed the random spy disguise script to have different presets (slow,
norm, fast) to correspond with the different movement speeds of the
disguises. Fast is enabled by default so that the random disguise will
never slow you down. The disguises in that range are scout, medic, and
spy.
This is a response to the recent change in Spy's movement speed.