1.10
- added an extra noflame feautre to pyro that hides the actual flames
when firing to improve visiblity.
- I SWITCHED TO SHIFT CROUCH! This is important. If you don't like it,
go to "STANDARD SETTINGS" in the autoexec and change "shiftcrouch" to
"ctrlcrouch".
- changed the pootis loop to VO loop to better support all voice
commands. Because it would've been the exact same script just with a
different voice command if I wanted to add other VO loops. Like "Help!"
or "Medic!".
- added a few more caption to prevent error messages in console. Purely
cosmetic to avoid red lines. It actually does nothing.
- added extra settings to the demo render config. This compensates for
the extra time it takes to render so chat and killstreak notices don't
just show up for half a second. Actual timing will vary on a machine and
case basis.